% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{builder}
\alias{builder}
\title{Building the clif matrix}
\usage{
builder(
  X,
  ind,
  real_pairs,
  exp_data,
  ligand_cell_list,
  ligand_index,
  recepts_index,
  receptor_cell_list,
  gene_name_index,
  cal_pos_coldirection_vec,
  cell_names
)
}
\arguments{
\item{X}{an initialized clif matrix}

\item{ind}{integer
index of a certain ligand-receptor pair}

\item{real_pairs}{character vector
a vector of ligand-receptor pairs}

\item{exp_data}{the scRNA-seq expression matrix}

\item{ligand_cell_list}{integer list
a list of cell indices that express the ligands}

\item{ligand_index}{integer vector
vector of the indices of the ligands}

\item{recepts_index}{integer vector
vector of the indices of the receptors}

\item{receptor_cell_list}{integer list
a list of cell indices that express the receptors}

\item{gene_name_index}{integer vector
vector of the indices of the genes}

\item{cal_pos_coldirection_vec}{function
calculating column positions in a vectorized way}

\item{cell_names}{character vector
vector of the (downsampled cell names)}
}
\description{
building the clif matrix where rows are
ligand-recetor pairs, columns are cell-pairs,
entries are the expression level
}
