% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{cal_pos_coldirection}
\alias{cal_pos_coldirection}
\title{Calculating column positions}
\usage{
cal_pos_coldirection(x, y, n)
}
\arguments{
\item{x}{integer
the index of the cell expressing the ligand gene}

\item{y}{integer
the index of the cell expressing the receptor gene}

\item{n}{integer
the total number of (downsampled) cells}
}
\value{
the column position of the cell-pair
}
\description{
Calculating the column positions of the cell-pair in the clif matrix
based on their indices
}
