% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{cal_pos_coldirection_vec}
\alias{cal_pos_coldirection_vec}
\title{Calculating column positions in a vectorized way}
\usage{
cal_pos_coldirection_vec(x, y, n)
}
\arguments{
\item{x}{integer vector
the indices of the cells expressing the ligand gene}

\item{y}{integer vector
the indices of the cells expressing the receptor gene}

\item{n}{integer
the total number of (downsampled) cells}
}
\value{
a vector of the column positions of the cell-pairs
}
\description{
Calculating the column positions of the cell-pair in the clif matrix
in a vectorized way based on their indices
}
