% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scc_seurat_functions.R
\name{FindAllMarkers}
\alias{FindAllMarkers}
\title{FindAllMarkers}
\usage{
FindAllMarkers(object, organization, ...)
}
\arguments{
\item{object}{an SCC object}

\item{organization}{the organization to run on}

\item{...}{parameters passed to base function in Seurat}
}
\value{
an SCC object
}
\description{
FindAllMarkers
}
