% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scc_seurat_functions.R
\name{FindClusters}
\alias{FindClusters}
\title{FindClusters}
\usage{
FindClusters(object, organization, ...)
}
\arguments{
\item{object}{an SCC object}

\item{organization}{the organization to run on}

\item{...}{parameters passed to base function in Seurat}
}
\value{
an SCC object
}
\description{
FindClusters
}
