% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scc_seurat_functions.R
\name{FindVariableFeatures}
\alias{FindVariableFeatures}
\title{FindVariableFeatures for SCC}
\usage{
FindVariableFeatures(object, organization, ...)
}
\arguments{
\item{object}{an SCC object}

\item{organization}{the organization to run on}

\item{...}{parameters passed to base function in Seurat}
}
\value{
an SCC object
}
\description{
FindVariableFeatures for SCC
}
