% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{geo_downsample}
\alias{geo_downsample}
\title{Geosketch downsampling}
\usage{
geo_downsample(exp_mat, sample_size_n = 1000)
}
\arguments{
\item{exp_mat, }{scRNA-seq expression matrix}

\item{sample_size_n}{integer, optional, default: 1000
the number of sampled cells if do downsampling}
}
\value{
an index of the sampled cells
}
\description{
Downsample the cells using a published method
called Geometric Sketching (details see Hie, Brian, et al.
 "Geometric sketching compactly summarizes the
 single-cell transcriptomic landscape." Cell systems 8.6 (2019): 483-493.)
}
