% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{load_lr_anno}
\alias{load_lr_anno}
\title{Load ligand-receptor annotations}
\usage{
load_lr_anno(species = "human")
}
\arguments{
\item{species}{character, default: "human"
accepted species are "human","mouse","rat","pig"}
}
\value{
a dataframe containing ligand-receptor from fantom5
}
\description{
Load ligand-receptor annotations for different species
}
