% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LoadCustom.R
\name{LoadCustom}
\alias{LoadCustom}
\title{Load the custom Ligand Receptor database from the user}
\usage{
LoadCustom(custom_LR_database)
}
\arguments{
\item{custom_LR_database}{data.frame. Each row is a ligand-receptor mechanism where the first column corresponds to the source genes that express the ligands subunits (separated by '_') and the second column corresponds to the receptor genes that express the receptor subunits (separated by '_').}
}
\description{
Load the custom Ligand Receptor database from the user
}
