% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LoadFantom5.R
\name{LoadFantom5}
\alias{LoadFantom5}
\title{Load FANTOM5 as ground truth}
\usage{
LoadFantom5(species)
}
\arguments{
\item{species}{species reference in the fantom 5 database}
}
\description{
Load FANTOM5 as ground truth
}
