% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LoadOmniPath.R
\name{LoadOmniPath}
\alias{LoadOmniPath}
\title{Load OmniPath as ground truth}
\usage{
LoadOmniPath(species)
}
\arguments{
\item{species}{Species reference in the Ominipath database}
}
\description{
Load OmniPath as ground truth
}
