% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ncomms8866}
\alias{ncomms8866}
\title{fantom5}
\format{
dataframe
}
\source{
fantom 5 database
}
\usage{
ncomms8866
}
\description{
Ligand-Receptor annotions from
fantom 5 database
}
\keyword{datasets}
