% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ncomms8866_human}
\alias{ncomms8866_human}
\title{fantom5 human}
\format{
dataframe
}
\source{
fantom 5 database
}
\usage{
ncomms8866_human
}
\description{
Human Ligand-Receptor annotions from
fantom 5 database
}
\keyword{datasets}
