% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ncomms8866_mouse}
\alias{ncomms8866_mouse}
\title{fantom5 mouse}
\format{
dataframe
}
\source{
fantom 5 database
}
\usage{
ncomms8866_mouse
}
\description{
Mouse Ligand-Receptor annotions from
fantom 5 database
}
\keyword{datasets}
