% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ncomms8866_pig}
\alias{ncomms8866_pig}
\title{fantom5 pig}
\format{
dataframe
}
\source{
fantom 5 database
}
\usage{
ncomms8866_pig
}
\description{
Pig Ligand-Receptor annotions from
fantom 5 database
}
\keyword{datasets}
