% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ncomms8866_rat}
\alias{ncomms8866_rat}
\title{fantom5 rat}
\format{
dataframe
}
\source{
fantom 5 database
}
\usage{
ncomms8866_rat
}
\description{
Rat Ligand-Receptor annotions from
fantom 5 database
}
\keyword{datasets}
