% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{prepSeurat}
\alias{prepSeurat}
\title{Seurat input preprocessing}
\usage{
prepSeurat(object, assay, min.cells.per.ident)
}
\arguments{
\item{object}{input Seurat object}

\item{assay}{which assay to use in object}

\item{min.cells.per.ident}{minimum cells per ident}
}
\value{
a processed seurat object
}
\description{
Seurat input preprocessing
}
