% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/return_celltypes.R
\name{return_celltypes}
\alias{return_celltypes}
\title{return cell types}
\usage{
return_celltypes(seurat_object)
}
\arguments{
\item{seurat_object}{Input seurat object}
}
\value{
unique cell types
}
\description{
return cell types
}
