% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReturnCellTypes.R
\name{ReturnCellTypes}
\alias{ReturnCellTypes}
\title{return cell types}
\usage{
ReturnCellTypes(seurat_object)
}
\arguments{
\item{seurat_object}{Input seurat object}
}
\value{
unique cell types
}
\description{
return cell types
}
