% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RunCellToCellSpatial.R
\name{RunCellToCellSpatial}
\alias{RunCellToCellSpatial}
\title{RunCellToCellSpatial}
\usage{
RunCellToCellSpatial(
  filtered.obj,
  ground.truth,
  assay,
  meta.data.to.map,
  edgelist,
  output_format
)
}
\arguments{
\item{filtered.obj}{A filtered Seurat object. The active identity will be used to define populations for connectomic sampling and crossings.}

\item{ground.truth}{Ground truth signaling mechanisms present in filtered.obj.}

\item{assay}{The assay to run the SCC transformation on. Defaults to "RNA."}

\item{meta.data.to.map}{A character vector of metadata names present in the original object which will be carried to the NICHES objects}

\item{edgelist}{data.frame. Each row is an directional edge between two spatially connected cells}

\item{output_format}{string. Choice of the output format. "seurat" will output a list of seurat objects, "raw" will output a list of lists with raw interaction matrix and compiled metadata}
}
\description{
RunCellToCellSpatial
}
