% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RunCellToSystem.R
\name{RunCellToSystem}
\alias{RunCellToSystem}
\title{RunCellToSystem}
\usage{
RunCellToSystem(
  object,
  LR.database,
  species,
  assay,
  min.cells.per.ident,
  min.cells.per.gene,
  blend = "sum",
  meta.data.to.map,
  ...
)
}
\arguments{
\item{object}{A Seurat 3.0 object.  The active identity meta.data will be used to define populations for connectomic sampling and crossings.}

\item{LR.database}{Currently accepts 'fantom5' or 'omnipath'}

\item{species}{The species of the object that is being processed.  Only required if LR.database = 'fantom5', and allows 'human','mouse','rat', or 'pig'}

\item{assay}{The assay to run the CellToSystem transformation on. Defaults to "RNA."}

\item{min.cells.per.ident}{A limit on how small (how many cells) a single population can be to participate in connectomic crossings.}

\item{min.cells.per.gene}{Limits analysis to interactions involving genes expressed above minimum threshold number of cells in the system.}

\item{blend}{Choice of linear operator to combine edges. Defaults to "sum", also accepts "mean"}

\item{meta.data.to.map}{A character vector of metadata names present in the original object which will be carried to the NICHES objects}

\item{...}{}
}
\description{
Condenses all signaling edges coming from each cell within a Seurat object connecting to every other cell in the system, treated as a single sink.
Outputs another Seurat object, but where the rows of the matrix are ligand-receptor mechanisms
and the columns are each a single sending cell barcode. The information in the matrix is a sum (or an average, depending on user preference) of
all signaling edges coming from that particular cell, to every other cell in the system (including to itself.)
This transformation allows rapid manipulation and dimensional reduction of how a cell is connected within the system.
The default assay of this object is called "CellToSystem" to distinguish it from other Seurat objects.
Meta.data slots by default contain "SendingType" information, which is the celltypes for each point, 
and "SendingCell" which is the exact cell barcode present in the original Seurat object.
}
