% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RunNICHES.R
\name{RunNICHES.matrix}
\alias{RunNICHES.matrix}
\title{RunNICHES for matrix}
\usage{
\method{RunNICHES}{matrix}(
  object,
  LR.database = "fantom5",
  species,
  min.cells.per.ident = NULL,
  min.cells.per.gene = NULL,
  meta.data.df = NULL,
  position.x = NULL,
  position.y = NULL,
  cell_types = NULL,
  custom_LR_database = NULL,
  k = 4,
  rad.set = NULL,
  blend = "mean",
  CellToCell = T,
  CellToSystem = F,
  SystemToCell = F,
  CellToCellSpatial = F,
  CellToNeighborhood = F,
  NeighborhoodToCell = F,
  output_format = "seurat"
)
}
\arguments{
\item{meta.data.df}{A dataframe. Optional. Default: NULL. A dataframe of the metadata (columns) associated with the cells (rows).}

\item{cell_types}{string. Optional. Default: NULL. The name of the meta.data.df column specifying the cell types of the cells. Only required for RunCellToCell.}
}
\description{
Performs user-selected NICHES transformations on a matrix. 
By default, references RunCellToCell, RunCellToSystem, and RunSystemToCell functions. 
If positional data is provided, similar analyses can be performed which are limited exclusively to cells that directly neighbor each other.
Output is a set of specialized NICHES objects which allow fine analysis of cell-cell interactions.
}
