% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RunNICHES.R
\name{RunNICHES.Seurat}
\alias{RunNICHES.Seurat}
\title{RunNICHES}
\usage{
\method{RunNICHES}{Seurat}(
  object,
  assay = "RNA",
  LR.database = "fantom5",
  species,
  min.cells.per.ident = NULL,
  min.cells.per.gene = NULL,
  meta.data.to.map = NULL,
  position.x = NULL,
  position.y = NULL,
  custom_LR_database = NULL,
  k = 4,
  rad.set = NULL,
  blend = "mean",
  CellToCell = T,
  CellToSystem = F,
  SystemToCell = F,
  CellToCellSpatial = F,
  CellToNeighborhood = F,
  NeighborhoodToCell = F,
  output_format = "seurat"
)
}
\arguments{
\item{assay}{string. Default: "RNA". The assay to run the NICHES transformation on.}

\item{meta.data.to.map}{character vector. Optional. Default: NULL. A vector of metadata names present in the original object which will be carried to the NICHES objects}
}
\description{
Performs user-selected NICHES transformations on a Seurat object. 
By default, references RunCellToCell, RunCellToSystem, and RunSystemToCell functions. 
If positional data is provided, similar analyses can be performed which are limited exclusively to cells that directly neighbor each other.
Output is a set of specialized NICHES objects which allow fine analysis of cell-cell interactions.
}
