% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runscc.R
\name{runSCC}
\alias{runSCC}
\title{RunSCC}
\usage{
runSCC(
  seu_obj,
  organizations = c("pair", "pair_spatial", "niche_spatial"),
  assay = NULL,
  metadata = NULL,
  species = "human",
  downsampling = TRUE,
  sample_size_n = 1000,
  autocrine = TRUE,
  neighborhood_radius = 1,
  n_threads = 8
)
}
\arguments{
\item{seu_obj}{input Seurat Object}

\item{organizations}{character vector, default: c("pair","pair_spatial","niche_spatial")
The type of SCC organizations to build}

\item{assay}{character, default: NULL
The assay to use from the input Seurat Object, NULL leads to the default assay}

\item{metadata}{character vector, default: NULL
The metadata to aggregate: e.g. cell_types, positions}

\item{species}{character, default: "human"
the species of the input scRNA-seq data}

\item{downsampling}{logical, default: TRUE
whether to do dowsampling on the cells}

\item{sample_size_n}{integer, default: 1000
the number of sampled cells if do downsampling}

\item{autocrine}{logical, default:TRUE
whether to consider cell interactions within one cell type
(including interact with themselves),always TRUE when no
cell_types provided}

\item{neighborhood_radius}{float, default:1
The euclidean distance threshold to consider as neighbors}

\item{n_threads}{integer,optional, default:8
the number of threads to use when constructing the matrix}
}
\value{
an SCC object which includes the specified organizations
}
\description{
Single-Cell Connectome (SCC) is a framework
that builds cell-cell ligand-receptor interactoin profiles
based on (spatial) single-cell transcriptomic data.
}
\examples{

scc_object <- runSCC(seu_obj = spatial_data,organizations=c("pair","pair_spatial","niche_spatial"),
 assay='alra',metadata = c("cell_types","position","cluster"),neighborhood_radius = 3,
 species = "mouse",downsampling = FALSE,sample_size_n = 10,autocrine =TRUE,n_threads = 8)


}
