% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runscc.R
\docType{class}
\name{SCC-class}
\alias{SCC-class}
\title{An S4 class to represent SCC hierarchy.}
\description{
An S4 class to represent SCC hierarchy.
}
\section{Slots}{

\describe{
\item{\code{organizations}}{A list containing different organizations}
}}

