% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base.R
\name{scc_base.default}
\alias{scc_base.default}
\title{scc_base default}
\usage{
\method{scc_base}{default}(
  exp_mat,
  species = "human",
  cell_types_vec = NULL,
  metadata_lst = NULL,
  downsampling = TRUE,
  sample_size_n = 1000,
  autocrine = TRUE,
  n_threads = 8
)
}
\arguments{
\item{cell_types_vec}{character vector, default: NULL
cell type annotations for each cell in the expressoin matrix, used for scc_base.Default}

\item{metadata_lst}{character list, optional, default: NULL
what metadata to aggregate, used for scc_base.Default}
}
\description{
scc_base default
}
