% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base.R
\name{scc_base}
\alias{scc_base}
\title{SCC base function}
\usage{
scc_base(exp_mat, ...)
}
\arguments{
\item{exp_mat}{scRNA-seq expression matrix}

\item{species}{character, default: "human"
the species of the input scRNA-seq data}

\item{downsampling}{logical, default: TRUE
whether to do dowsampling on the cells}

\item{sample_size_n}{integer, default: 1000
the number of sampled cells if do downsampling}

\item{autocrine}{logical,default:TRUE
whether to consider cell interactions within one cell type
(including interact with themselves),always TRUE is no
cell_types_vec provided}

\item{n_threads}{integer,default:8
the number of threads to use when constructing the matrix}
}
\value{
a Seurat object containing the clif matrix is returned
}
\description{
Construct the fully connected cell-pair ligand-receptor expression profiles
}
\examples{

}
