% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base.R
\name{scc_base.Seurat}
\alias{scc_base.Seurat}
\title{scc_base Seurat}
\usage{
\method{scc_base}{Seurat}(
  seu_obj,
  assay = NULL,
  metadata,
  species = "human",
  downsampling = TRUE,
  sample_size_n = 1000,
  autocrine = TRUE,
  n_threads = 8
)
}
\arguments{
\item{assay}{character, default: NULL
The assay to use from the input Seurat Object, NULL leads to the default assay, used for scc_base.Seurat}

\item{metadata}{character vector, default: NULL
The metadata to aggregate: e.g. cell_types, positions, used for scc_base.Seurat}
}
\description{
scc_base Seurat
}
