% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FilterGroundTruth.R
\name{FilterGroundTruth}
\alias{FilterGroundTruth}
\title{Filter the ground truth to only include mechanisms in the base object}
\usage{
FilterGroundTruth(ground.truth, input_rownames)
}
\arguments{
\item{ground.truth}{Unfiltered LR mechanisms from the database}

\item{input_rownames}{Gene names of the filtered seurat object}
}
\value{
Filtered ground truth signaling mechanisms.
}
\description{
Filter the ground truth to only include mechanisms in the base object
}
