% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSeuratAssay.R
\name{getSeuratAssay}
\alias{getSeuratAssay}
\title{Wrapper function for GetAssayData based on the version of the SeuratObject}
\usage{
getSeuratAssay(object, assay, slot)
}
\arguments{
\item{object}{input Seurat object}

\item{assay}{which assay to use in object}

\item{slot}{"data","counts",or "scale.data"}
}
\value{
the specified assay data
}
\description{
Wrapper function for GetAssayData based on the version of the SeuratObject
}
