% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{lr_load}
\alias{lr_load}
\title{load the Ligands and Receptors that corresponds to the row names of the input seurat}
\usage{
lr_load(LR.database, species, input_rownames)
}
\arguments{
\item{LR.database}{database to use}

\item{species}{}

\item{input_rownames}{the genes names to query}
}
\value{
a list of ligands and receptors
}
\description{
load the Ligands and Receptors that corresponds to the row names of the input seurat
}
