% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RunCellToCell.R
\name{RunCellToCell}
\alias{RunCellToCell}
\title{RunCellToCell}
\usage{
RunCellToCell(
  filtered.obj,
  ground.truth,
  assay,
  meta.data.to.map,
  output_format
)
}
\arguments{
\item{filtered.obj}{A filtered Seurat object. The active identity will be used to define populations for connectomic sampling and crossings.}

\item{ground.truth}{Ground truth signaling mechanisms present in filtered.obj.}

\item{assay}{The assay to run the SCC transformation on. Defaults to "RNA."}

\item{meta.data.to.map}{A character vector of metadata names present in the original object which will be carried to the SCC objects}

\item{output_format}{string. Choice of the output format. "seurat" will output a list of seurat objects, "raw" will output a list of lists with raw interaction matrix and compiled metadata}
}
\description{
Performs cell-cell transformation on a Seurat object, with structural downsampling to avoid data-inflation. Outputs another Seurat object, but where the columns of the matrix are 
barcode-barcode pairs, and the rows of the matrix are ligand-receptor mechanisms. This allows rapid manipulation and dimensional reduction of cell-cell connectivity data.
The default assay of this object is called "CellToCell" to distinguish it from normal Seurat objects.
Meta.data slots by default contain "SendingType" "ReceivingType" and "VectorType" information.
}
