% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RunCellToCell.R
\name{RunCellToCell}
\alias{RunCellToCell}
\title{RunCellToCell}
\usage{
RunCellToCell(
  object,
  LR.database,
  species,
  assay,
  min.cells.per.ident,
  meta.data.to.map,
  ...
)
}
\arguments{
\item{object}{A Seurat 3.0 object.  The active identity meta.data will be used to define populations for connectomic sampling and crossings.}

\item{LR.database}{Accepts either 'fantom5' or a custom data.frame with the first column equal to ligands, second column equal to associated receptors.}

\item{species}{The species of the object that is being processed.  Only required if LR.database = 'fantom5', and allows 'human','mouse','rat', or 'pig'}

\item{assay}{The assay to run the SCC transformation on. Defaults to "RNA."}

\item{min.cells.per.ident}{Default 1. A limit on how small (how many cells) a single population can be to participate in connectomic crossings.}

\item{meta.data.to.map}{A character vector of metadata names present in the original object which will be carried to the SCC objects}
}
\description{
Performs cell-cell transformation on a Seurat object, with structural downsampling to avoid data-inflation. Outputs another Seurat object, but where the columns of the matrix are 
barcode-barcode pairs, and the rows of the matrix are ligand-receptor mechanisms. This allows rapid manipulation and dimensional reduction of cell-cell connectivity data.
The default assay of this object is called "CellToCell" to distinguish it from normal Seurat objects.
Meta.data slots by default contain "SendingType" "ReceivingType" and "VectorType" information.
}
