% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RunCellToNeighborhood.R
\name{RunCellToNeighborhood}
\alias{RunCellToNeighborhood}
\title{RunCellToNeighborhood}
\usage{
RunCellToNeighborhood(
  sys.small,
  ground.truth,
  assay,
  position.x,
  position.y,
  rad.set,
  meta.data.to.map,
  ...
)
}
\arguments{
\item{sys.small}{A filtered Seurat object. The active identity will be used to define populations for connectomic sampling and crossings.}

\item{ground.truth}{Ground truth signaling mechanisms present in sys.small.}

\item{assay}{The assay to run the SCC transformation on. Defaults to "RNA."}

\item{position.x}{The name of the meta.data column specifying location on the spatial x-axis. Only relevant for spatial omics data.}

\item{position.y}{The name of the meta.data column specifying location on the spatial y-axis. Only relevant for spatial omics data.}

\item{rad.set}{The radius in Euclidean space to consider local neighbors.}

\item{meta.data.to.map}{A character vector of metadata names present in the original object which will be carried to the NICHES objects}

\item{...}{}
}
\description{
RunCellToNeighborhood
}
