% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RunCellToSystem.R
\name{RunCellToSystem}
\alias{RunCellToSystem}
\title{RunCellToSystem}
\usage{
RunCellToSystem(
  object,
  LR.database = "fantom5",
  species,
  assay = "RNA",
  min.cells.per.ident = 1,
  blend = "sum"
)
}
\arguments{
\item{object}{A Seurat 3.0 object.  The active identity meta.data will be used to define populations for connectomic sampling and crossings.}

\item{LR.database}{Accepts either 'fantom5' or a custom data.frame with the first column equal to ligands, second column equal to associated receptors.}

\item{species}{The species of the object that is being processed.  Only required if LR.database = 'fantom5', and allows 'human','mouse','rat', or 'pig'}

\item{assay}{The assay to run the CellToSystem transformation on. Defaults to "RNA."}

\item{min.cells.per.ident}{Default 1. A limit on how small (how many cells) a single population can be to participate in connectomic crossings.}

\item{blend}{Choice of linear operator to combine edges. Defaults to "sum", also accepts "mean"}
}
\description{
Condenses all signaling edges coming from each cell within a Seurat object connecting to every other cell in the system, including itself. Outputs another Seurat object, but where the rows of the matrix are ligand-receptor mechanisms
and the columns are each a single cell barcode. The information in the matrix is a sum (or an average, depending on user preference) of
all signaling edges coming from that particular cell, to every other cell in the system (including to itself.)
This transformation allows rapid manipulation and dimensional reduction of how a cell is connected within the system.
The default assay of this object is called "CellToSystem" to distinguish it from other Seurat objects.
Meta.data slots by default contain "SendingType" information, which is the celltypes for each point, 
and "SendingCell" which is the exact cell barcode present in the original Seurat object
}
