% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RunNeighborhoodToCell.R
\name{RunNeighborhoodToCell}
\alias{RunNeighborhoodToCell}
\title{RunNeighborhoodToCell}
\usage{
RunNeighborhoodToCell(
  object,
  LR.database = "fantom5",
  species,
  assay = "RNA",
  min.cells.per.ident = 1,
  position.x,
  position.y,
  meta.data.to.map = NULL
)
}
\arguments{
\item{object}{A Seurat 4.0 object. The active identity will be used to define populations for connectomic sampling and crossings.}

\item{LR.database}{Accepts either 'fantom5' or a custom data.frame with the first column equal to ligands, second column equal to associated receptors.}

\item{species}{The species of the object that is being processed.  Only required if LR.database = 'fantom5', and allows 'human','mouse','rat', or 'pig'}

\item{assay}{The assay to run the SCC transformation on. Defaults to "RNA."}

\item{min.cells.per.ident}{Default 1. A limit on how small (how many cells) a single population can be to participate in connectomic crossings.}

\item{position.x}{The name of the meta.data column specifying location on the spatial x-axis. Only relevant for spatial omics data.}

\item{position.y}{The name of the meta.data column specifying location on the spatial y-axis. Only relevant for spatial omics data.}
}
\description{
RunNeighborhoodToCell
}
