% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RunNeighborhoodToCell.R
\name{RunNeighborhoodToCell}
\alias{RunNeighborhoodToCell}
\title{RunNeighborhoodToCell}
\usage{
RunNeighborhoodToCell(
  object,
  LR.database,
  species,
  assay,
  min.cells.per.ident,
  min.cells.per.gene,
  position.x,
  position.y,
  rad.set,
  meta.data.to.map,
  ...
)
}
\arguments{
\item{object}{A Seurat 4.0 object. The active identity will be used to define populations for connectomic sampling and crossings.}

\item{LR.database}{Currently accepts 'fantom5' or 'omnipath'}

\item{species}{The species of the object that is being processed.  Only required if LR.database = 'fantom5', and allows 'human','mouse','rat', or 'pig'}

\item{assay}{The assay to run the SCC transformation on. Defaults to "RNA."}

\item{min.cells.per.ident}{A limit on how small (how many cells) a single population can be to participate in connectomic crossings.}

\item{min.cells.per.gene}{Limits analysis to interactions involving genes expressed above minimum threshold number of cells in the system.}

\item{position.x}{The name of the meta.data column specifying location on the spatial x-axis. Only relevant for spatial omics data.}

\item{position.y}{The name of the meta.data column specifying location on the spatial y-axis. Only relevant for spatial omics data.}

\item{rad.set}{The radius in Euclidean space to consider local neighbors.}

\item{meta.data.to.map}{A character vector of metadata names present in the original object which will be carried to the NICHES objects}

\item{...}{}
}
\description{
RunNeighborhoodToCell
}
