% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RunNICHES.R
\name{RunNICHES}
\alias{RunNICHES}
\title{RunNICHES}
\usage{
RunNICHES(
  object,
  LR.database,
  species,
  assay,
  min.cells.per.ident = 1,
  min.cells.per.gene = 10,
  meta.data.to.map = NULL,
  position.x = NULL,
  position.y = NULL,
  rad.set = 1,
  blend = "sum",
  CellToCell = T,
  CellToSystem = F,
  SystemToCell = F,
  CellToCellSpatial = F,
  CellToNeighborhood = F,
  NeighborhoodToCell = F,
  ...
)
}
\arguments{
\item{object}{A Seurat 4.0 object. The active identity will be used to define populations for connectomic sampling and crossings.}

\item{LR.database}{Currently accepts 'fantom5' or 'omnipath'}

\item{species}{The species of the object that is being processed. Only required if LR.database = 'fantom5', and allows 'human','mouse','rat', or 'pig'}

\item{assay}{The assay to run the NICHES transformation on. Defaults to "RNA."}

\item{min.cells.per.ident}{Default 1. A limit on how small (how many cells) a single population can be to participate in connectomic crossings.}

\item{min.cells.per.gene}{Default 10. Limits analysis to interactions involving genes expressed above minimum threshold number of cells in the system.}

\item{meta.data.to.map}{A character vector of metadata names present in the original object which will be carried to the NICHES objects}

\item{position.x}{The name of the meta.data column specifying location on the spatial x-axis. Only relevant for spatial omics data.}

\item{position.y}{The name of the meta.data column specifying location on the spatial y-axis. Only relevant for spatial omics data.}

\item{rad.set}{Default 1. The radius in Euclidean space to consider local neighbors.}

\item{blend}{Choice of linear operator to combine edges in single-cell niche investigations. Defaults to "sum", also accepts "mean".}

\item{CellToCell}{Default TRUE. Whether to analyze cell-cell interactions without considering spatial coordinates.}

\item{CellToSystem}{Default FALSE. Whether to analyze summed signaling output to total system coming from each cell. Does not consider Euclidean coordinates.}

\item{SystemToCell}{Default FALSE. Whether to analyze summed signaling input from total system landing on each cell (cellular microenvironment/niche). Does not consider Euclidean coordinates.}

\item{CellToCellSpatial}{Default FALSE. Whether to analyze cell-cell interactions between Euclidean neighbors. Only applicable in spatial datasets.}

\item{CellToNeighborhood}{Default FALSE. Whether to analyze summed signaling output to Euclidean neighbors. Only applicable in spatial datasets.}

\item{NeighborhoodToCell}{Default FALSE. Whether to analyze summed signaling input from Euclidean neighbors (cellular microenvironment/niche). Only applicable in spatial datasets.}

\item{...}{Additional parameters to pass to RunCellToCell, RunSystemToCell, RunCellToSystem, or spatial equivalents}
}
\description{
Performs user-selected NICHES transformations on a Seurat object. 
By default, references RunCellToCell, RunCellToSystem, and RunSystemToCell functions. 
If positional data is provided, similar analyses can be performed which are limited exclusively to cells that directly neighbor each other.
Output is a set of specialized NICHES objects which allow fine analysis of cell-cell interactions.
}
