% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RunNICHES.R
\name{RunNICHES}
\alias{RunNICHES}
\title{RunNICHES}
\usage{
RunNICHES(object, ...)
}
\arguments{
\item{object}{normalized data matrix or a seurat object}

\item{...}{}
}
\description{
RunNICHES generic function
}
