% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RunSCC.R
\name{RunSCC}
\alias{RunSCC}
\title{RunSCC}
\usage{
RunSCC(
  object,
  LR.database = "fantom5",
  species,
  assay = "RNA",
  min.cells.per.ident = 10,
  meta.data.to.map = NULL,
  position.x = NULL,
  position.y = NULL,
  CellCell = T,
  CellSystem = T,
  SystemCell = T,
  CellCellSpatial = T,
  CellNeighborhood = T,
  NeighborhoodCell = T,
  ...
)
}
\arguments{
\item{object}{A Seurat 4.0 object. The active identity will be used to define populations for connectomic sampling and crossings.}

\item{LR.database}{Accepts either 'fantom5' or a custom data.frame with the first column equal to ligands, second column equal to associated receptors.}

\item{species}{The species of the object that is being processed. Only required if LR.database = 'fantom5', and allows 'human','mouse','rat', or 'pig'}

\item{assay}{The assay to run the SCC transformation on. Defaults to "RNA."}

\item{min.cells.per.ident}{Default 10. A limit on how small (how many cells) a single population can be to participate in connectomic crossings.}

\item{meta.data.to.map}{A character vector of metadata names present in the original object which will be carried to the SCC objects}

\item{position.x}{The name of the meta.data column specifying location on the spatial x-axis. Only relevant for spatial omics data.}

\item{position.y}{The name of the meta.data column specifying location on the spatial y-axis. Only relevant for spatial omics data.}

\item{...}{Additional parameters to pass to RunCellCell, RunSystemCell, RunCellSystem, or spatial equivalents}
}
\description{
Performs Single-Cell Connectivity (SCC) transformations on a Seurat object. 
By default, references RunCellCell, RunCellSystem, and RunSystemCell functions. 
If positional data is provided, similar analyses can be performed which are limited exclusively to cells that directly neighbor each other.
Output is a set of specialized Seurat objects (SCC objects) corresponding to input arguments.
}
