% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RunSystemToCell.R
\name{RunSystemToCell}
\alias{RunSystemToCell}
\title{RunSystemToCell}
\usage{
RunSystemToCell(
  object,
  LR.database,
  species,
  assay,
  min.cells.per.ident,
  blend = "sum",
  meta.data.to.map,
  ...
)
}
\arguments{
\item{object}{A Seurat 3.0 object.  The active identity meta.data will be used to define populations for connectomic sampling and crossings.}

\item{LR.database}{Currently accepts 'fantom5' or 'omnipath'}

\item{species}{The species of the object that is being processed.  Only required if LR.database = 'fantom5', and allows 'human','mouse','rat', or 'pig'}

\item{assay}{The assay to run the SystemToCell transformation on. Defaults to "RNA."}

\item{min.cells.per.ident}{Default 1. A limit on how small (how many cells) a single population can be to participate in connectomic crossings.}

\item{blend}{Choice of linear operator to combine edges. Defaults to "sum", also accepts "mean"}

\item{meta.data.to.map}{A character vector of metadata names present in the original object which will be carried to the NICHES objects}

\item{...}{}
}
\description{
Condenses signaling edges landing on each cell within a Seurat object, from all other cells in the system treated as a single source. 
Outputs another Seurat object, but where the rows of the matrix are ligand-receptor mechanisms
and the columns are each a single receiving cell barcode. The information in the matrix is a sum (or an average, depending on user preference) of
all signaling edges landing on that particular cell, from all cells in the system (including from itself.)
This transformation allows rapid manipulation and dimensional reduction of how a cell is connected within the system.
The default assay of this object is called "SystemToCell" to distinguish it from other Seurat objects.
Meta.data slots by default contain "ReceivingType" information, which is the celltypes for each point, 
and "ReceivingCell" which is the exact cell barcode present in the original Seurat object.
}
